+++ src/app/sidebar/Sidebar.stories.tsx
@@ -2,7 +2,7 @@ import { Story, Meta } from '@storybook/react/types-6-0';
 import React from 'react';
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 
-import { Sidebar, SidebarProps } from './Sidebar';
+import { Sidebar, SidebarProps } from './components/Sidebar';
 
 export default {
   title: 'Sidebar',
