+++ src/app/sidebar/index.tsx
@@ -1,0 +1,1 @@
+export { SidebarContainer as Sidebar } from './containers/SidebarContainer';
