+++ src/pages/calendar/index.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from 'react';
 import { useNavigate, useParams } from 'react-router';
 import { useSearchParams } from 'react-router-dom';
 
-import { Header, Content, SidebarContainer, Feedback } from '../../app/index';
+import { Header, Content as CourseContent, SidebarContainer, Feedback } from '../../app/index';
 import { getCurrentTerm } from '../../app/shared/utils/terms';
 import { Term } from '../../shared/fetchers';
 
@@ -40,7 +40,7 @@ export function Calendar(): JSX.Element {
         />
         <Flex minW="80%" overflow="auto" justifyContent="center" height="100%" boxShadow="lg" zIndex={56}>
           {term && subject && code && searchParams.get('pid') ? (
-            <Content term={term as Term} />
+            <CourseContent term={term as Term} />
           ) : (
             <Center p="10">
               <VStack>
